added footer component; moved Navigation cmpt/scss to Navigation folder; moved assets folder to src folder
