+++ src/assets/Logo/InStock-Logo_2x.png
