+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import "./Footer.scss";
+
+const Footer = () => {
+  return (
+    <div className="footer">
+      <p className="footer__header">&copy; InStock Inc. All Rights Reserved.</p>
+    </div>
+  );
+};
+
+export default Footer;
