+++ src/components/Footer/Footer.scss
@@ -1,0 +1,13 @@
+@use "../../styles/partials/fonts" as *;
+@use "../../styles/partials/variables" as *;
+
+
+.footer {
+    padding: 36px;
+
+    .footer__header {
+        color: $instock-slate;
+        font-family: 'Titillium Web', sans-serif;
+        text-align: center;
+    }
+}
