+++ src/components/Navigation/Navigation.scss
@@ -1,6 +1,6 @@
-@use "../styles/partials/fonts" as *;
-@use "../styles/partials/variables" as variables;
-@use "../styles/partials/mixins" as *;
+@use "../../styles/partials/fonts" as *;
+@use "../../styles/partials/variables" as variables;
+@use "../../styles/partials/mixins" as *;
 
 nav {
   background-color: variables.$instock-graphite;
